requirements.txt
Flask==3.0.3
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.0.1
marshmallow==3.22.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.25.1
python-dotenv==1.0.1
PyJWT==2.8.0
